+++ docs/src/pages/system/global-classes/Basics.js
@@ -1,33 +1,5 @@
-import { withStyles } from '@material-ui/styles';
-
 import Button from '@material-ui/core/Button';
-import spacings from './spacings';
-import colors from './colors';
-import elevations from './elevations';
-import texts from './texts';
-import {
-  displays,
-  overflows,
-  textOverflows,
-  visibilities,
-  whiteSpaces,
-} from './displays';
-
-const GlobalCss = withStyles((theme) => {
-  return {
-    '@global': {
-      ...spacings(theme),
-      ...colors(theme),
-      ...elevations(theme),
-      ...texts(theme),
-      ...displays,
-      ...overflows,
-      ...textOverflows,
-      ...visibilities,
-      ...whiteSpaces,
-    },
-  };
-})(() => null);
+import GlobalCss from './GlobalCss';
 
 export default function App() {
   return (
