+++ docs/src/pages/system/global-classes/Basics.tsx
@@ -1,33 +1,5 @@
-import { withStyles } from '@material-ui/styles';
-import { Theme } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import spacings from './spacings';
-import colors from './colors';
-import elevations from './elevations';
-import texts from './texts';
-import {
-  displays,
-  overflows,
-  textOverflows,
-  visibilities,
-  whiteSpaces,
-} from './displays';
-
-const GlobalCss = withStyles((theme: Theme) => {
-  return {
-    '@global': {
-      ...spacings(theme),
-      ...colors(theme),
-      ...elevations(theme),
-      ...texts(theme),
-      ...displays,
-      ...overflows,
-      ...textOverflows,
-      ...visibilities,
-      ...whiteSpaces,
-    },
-  };
-})(() => null);
+import GlobalCss from './GlobalCss';
 
 export default function App() {
   return (
