+++ docs/src/pages/system/global-classes/GlobalCss.d.ts
@@ -1,0 +1,1 @@
+export default function (): JSX.Element;
